+++ src/components/CardCompanySelect/CardCompanySelect.tsx
@@ -2,17 +2,9 @@ import { useState } from "react";
 import styles from "./CardCompanySelect.module.css";
 import InputContainer from "../InputContainer/InputContainer";
 import { useCardContext } from "../../contexts/CardContext";
+import { INPUT_CONTAINER } from "../../constants/title";
+import { CARD_COMPANIES } from "../../constants/cardCompanyInfo";
 
-export const CARD_COMPANIES = [
-  { name: "BC카드", color: "#F04651" },
-  { name: "신한카드", color: "#0046FF" },
-  { name: "카카오뱅크", color: "#FFE600" },
-  { name: "현대카드", color: "#000000" },
-  { name: "우리카드", color: "#007BC8" },
-  { name: "롯데카드", color: "#ED1C24" },
-  { name: "하나카드", color: "#009490" },
-  { name: "국민카드", color: "#6A6056" },
-];
 const CARD_OPTIONS = CARD_COMPANIES.map((company) => company.name);
 
 const CardCompanySelect = () => {
@@ -33,13 +25,13 @@ const CardCompanySelect = () => {
 
   return (
     <InputContainer
-      title="카드사를 선택해 주세요"
-      subTitle="현재 국내 카드사만 가능합니다."
+      title={INPUT_CONTAINER.CARD_COMPANY.TITLE}
+      subTitle={INPUT_CONTAINER.CARD_COMPANY.SUBTITLE}
     >
       <div className={styles.wrapper}>
         <div className={styles.selectBox} onClick={toggleOpen} tabIndex={0}>
           <span className={selected ? styles.selected : styles.placeholder}>
-            {selected || "카드사를 선택해주세요"}
+            {selected || INPUT_CONTAINER.CARD_COMPANY.PLACEHOLDER}
           </span>
           <span className={styles.arrow}>▾</span>
         </div>
