+++ src/constants/cardCompanyInfo.ts
@@ -1,0 +1,10 @@
+export const CARD_COMPANIES = [
+  { name: "BC카드", color: "#F04651" },
+  { name: "신한카드", color: "#0046FF" },
+  { name: "카카오뱅크", color: "#FFE600" },
+  { name: "현대카드", color: "#000000" },
+  { name: "우리카드", color: "#007BC8" },
+  { name: "롯데카드", color: "#ED1C24" },
+  { name: "하나카드", color: "#009490" },
+  { name: "국민카드", color: "#6A6056" },
+];
