+++ src/constants/title.ts
@@ -10,4 +10,13 @@ export const INPUT_CONTAINER = {
   CVC: {
     TITLE: "CVC 번호를 입력해 주세요",
   },
+  CARD_COMPANY: {
+    TITLE: "카드사를 선택해 주세요",
+    SUBTITLE: "현재 국내 카드사만 가능합니다.",
+    PLACEHOLDER: "카드사 선택",
+  },
+  PASSWORD: {
+    TITLE: "비밀번호를 입력해 주세요",
+    SUBTITLE: "앞의 2자리를 입력해주세요",
+  },
 };
